test: move fixtures to test, check for change password
